Separated server test functions
